refactor employer poc operations
