--- app/domain/entities/employer_staff_role.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Entities
-  class EmployerStaffRole < Dry::Struct
-    transform_keys(&:to_sym)
-
-    attribute :is_owner, Types::Bool
-    attribute :benefit_sponsor_employer_profile_id, Types::Bson
-    attribute :coverage_record, Entities::CoverageRecord
-  end
-end
